Move Window constants for use in overlay
