+++ src/main/view.ts
@@ -1,7 +1,7 @@
 import { BrowserView, app, ContextMenuParams } from "electron";
 import { resolve } from "path";
 import { appWindow } from ".";
-import { NAVIGATION_HEIGHT } from "../renderer/app/constants/window";
+import { NAVIGATION_HEIGHT } from "../renderer/constants/window";
 import { getGeneralMenu } from "./menus/general";
 import { downloadFaviconFromUrl } from "./tools/favicon";
 import { BLUE_1 } from "../renderer/constants/colors";
