+++ src/renderer/constants/window.ts
